+++ tests/lint/octave_global_functions/foo.m
@@ -1,0 +1,7 @@
+1;
+
+function x = potato(y)
+    x = y + 1;
+end
+
+disp('Hello world!');