+++ tests/lint/octave_global_functions/test_1.m
@@ -1,0 +1,6 @@
+foo;
+disp(potato(12));
+
+% Not an error, because foo is a bizarre octave extension that makes
+% functions in the middle go global. Specifically this happens only
+% for _script_ files