+++ tests/lint/octave_global_functions/test_1_b.m
@@ -1,0 +1,3 @@
+disp(potato(12));
+
+% Error, because potato (from the script file foo) is not visible.