+++ tests/lint/octave_global_functions/test_2.m
@@ -1,0 +1,3 @@
+disp(wibble(5));
+
+% Should print 11.