+++ tests/lint/octave_global_functions/test_3.m
@@ -1,0 +1,3 @@
+disp(wobble(5));
+
+% Error, because wobble is not visible/private