+++ tests/lint/octave_global_functions/test_4.m
@@ -1,0 +1,5 @@
+disp(wibble(5));
+% Prints 11
+
+disp(wobble(5));
+% Error, because wobble is still not visible/private