+++ tests/lint/octave_global_functions/wibble.m
@@ -1,0 +1,7 @@
+function rv = wibble(x)
+    rv = wobble(x) + 1;
+end
+
+function rv = wobble(x)
+    rv = x * 2;
+end